Update README.md
Scope of the Proposed System: - 
 
The main objective is to develop a “ECommerce website” that covers all the aspects of shopping online. 
 
Goals  
 
The primary goal of e-commerce is to reach maximum customers at 
the right time to increase sales and profitability of the business. 
Flexibility for customers. 
Product and price comparison. User friendly. 
 
Objectives  
 
•	To make various products available to different citizens across India at best rate. 
•	As time flow by, technology is evolving and most people love to do things with least labor. Like booking a cab with an app. Through this website people can shop huge number of things without stepping outside and can pay on delivery or online. 
•	To provide a smooth interface to the user and help him/her find their desired product easily. 
 
Users 
 
Contact Persons (Admin): 
These types of users perform all the transactions on behalf of the company. They can place orders on the company account and billing address. The admin can view orders of other contact persons of the same company. The admin can add or delete other contact persons of the company as well. 
Contact Persons (Non-Admin): 
The same as the contact person admin, the only difference is that they neither can create other contact persons, nor they, can view orders places by other